+++ examples/AutoBenchmark/generate_readme.py
@@ -77,43 +77,54 @@ The CPU times below are given in microseconds.
 
 ## CPU Time Changes
 
-* The CPU time did not change much from v0.8 to v1.4.
-* The CPU time of most classes did not change much from v1.4 to v1.5. The
-  big difference is that the Zone registries (kZoneRegistry,
-  kZoneAndLinkRegistry) are now sorted by zoneId instead of zoneName, and the
-  `ZoneManager::createForZoneId()` will use a binary search, instead of a linear
-  search. This makes it 10-15X faster for ~266 entries. The
-  `ZoneManager::createForZoneName()` also converts to a zoneId, then performs a
-  binary search, instead of doing a binary search on the zoneName directly. Even
-  with the extra level of indirection, the `createForZoneName()` is between
-  1.5-2X faster than the previous version.
-* In v1.6, BasicZoneManager and ExtendedZoneManager can take an optional
+v0.8 to v1.4:
+* The CPU time did not change much from
+
+In v1.5:
+* No significant changes to CPU time.
+* Zone registries (kZoneRegistry, kZoneAndLinkRegistry) are now sorted by zoneId
+  instead of zoneName, and the `ZoneManager::createForZoneId()` will use a
+  binary search, instead of a linear search. This makes it 10-15X faster for
+  ~266 entries.
+* The `ZoneManager::createForZoneName()` also converts to a zoneId, then
+  performs a binary search, instead of doing a binary search on the zoneName
+  directly. Even with the extra level of indirection, the `createForZoneName()`
+  is between 1.5-2X faster than the previous version.
+
+In v1.6:
+* BasicZoneManager and ExtendedZoneManager can take an optional
   LinkRegistry which will be searched if a zoneId is not found. The
   `BasicZoneManager::createForZoneId(link)` benchmark shows that if the zoneId
   is not found, the total search time is roughly double, because the
-  LinkRegistry must be search as a fallback. On some compilers, the
-  `BasicZoneManager::createForZoneName(binary)` becames slightly slower (~10%?)
-  because the algorithm was moved into the `ace_common::binarySearchByKey()`
-  template function, and the compiler is not able to optimize the resulting
-  function as well as the hand-rolled version. The slightly decrease in speed
-  seemed acceptable cost to reduce duplicate code maintenance.
-* In v1.7.2, `SystemClock::clockMillis()` became non-virtual after incorporating
+  LinkRegistry must be search as a fallback.
+* On some compilers, the `BasicZoneManager::createForZoneName(binary)` becames
+  slightly slower (~10%?) because the algorithm was moved into the
+  `ace_common::binarySearchByKey()` template function, and the compiler is not
+  able to optimize the resulting function as well as the hand-rolled version.
+  The slightly decrease in speed seemed acceptable cost to reduce duplicate code
+  maintenance.
+
+In v1.7.2:
+* `SystemClock::clockMillis()` became non-virtual after incorporating
   AceRoutine v1.3. The sizeof `SystemClockLoop` and `SystemClockCoroutine`
   decreases 4 bytes on AVR, and 4-8 bytes on 32-bit processors. No signficant
   changes in CPU time.
-* In v1.7.4+:
-    * changes to size of `ExtendedZoneProcessor`:
-        * 8-bit processors
-            * increases by 24 bytes on AVR, due adding 1 pointer and 2
-              `uint16_t` to MatchingEra
-            * decreases by 48 bytes on AVR, by disabling
-              `originalTransitionTime` unless
-              `ACE_TIME_EXTENDED_ZONE_PROCESSOR_DEBUG` is enabled.
-        * 32-bit processors
-            * increases by 32 bytes on 32-bit processors due to adding
-              a pointer and 2 `uint16_t` to MatchingEra
-            * decreases by 32 bytes on 32-bit processors due to disabling
-              `originalTransitionTime` in Transition
+
+In v1.7.4+:
+* significant changes to size of `ExtendedZoneProcessor`
+    * 8-bit processors
+        * increases by 24 bytes on AVR, due adding 1 pointer and 2
+            `uint16_t` to MatchingEra
+        * decreases by 48 bytes on AVR, by disabling
+            `originalTransitionTime` unless
+            `ACE_TIME_EXTENDED_ZONE_PROCESSOR_DEBUG` is enabled.
+    * 32-bit processors
+        * increases by 32 bytes on 32-bit processors due to adding
+            a pointer and 2 `uint16_t` to MatchingEra
+        * decreases by 32 bytes on 32-bit processors due to disabling
+            `originalTransitionTime` in Transition
+* Upgrade ESP8266 Core from 2.7.4 to 3.0.2.
+    * AutoBenchmark indicate that things are a few percentage faster.
 
 ## Arduino Nano
 
@@ -159,7 +170,7 @@ The CPU times below are given in microseconds.
 
 * NodeMCU 1.0 clone, 80MHz ESP8266
 * Arduino IDE 1.8.13
-* ESP8266 Boards 2.7.4
+* ESP8266 Boards 3.0.2
 
 ```
 {esp8266_results}